Remove httplib2 from requirements.txt
The version of `httplib2` specified in this requirements.txt has been flagged by GitHub's dependabot as a security vulnerability.  `httplib2` is a dependency of `google-auth-httplib2`, and not imported directly by any scripts in this repo, so it is not necessary to specify it explicitly here.  I have removed it accordingly.